+++ templates/haskell/nix/shell.nix
@@ -47,7 +47,7 @@ cabalProject:
     # shellcheck = null;
     # prettier = null;
     # editorconfig-checker = null;
-    # nixfmt = null;
+    # nixfmt-classic = null;
     # optipng = null;
     # purs-tidy = null;
   };
@@ -67,8 +67,8 @@ cabalProject:
   #   prettier.extraOptions = "";
   #   editorconfig-checker.enable = false;
   #   editorconfig-checker.extraOptions = "";
-  #   nixfmt.enable = false;
-  #   nixfmt.extraOptions = "";
+  #   nixfmt-classic.enable = false;
+  #   nixfmt-classic.extraOptions = "";
   #   optipng.enable = false;
   #   optipng.extraOptions = "";
   #   purs-tidy.enable = false;