+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 19 Nov 2024
 
-- Removed now-archived `nixpkgs-fmt` in favor of `nixfmt`. 
+- Removed now-archived `nixpkgs-fmt` in favor of `nixfmt-classic`. 
   
   To migrate, replace:
   ```
@@ -17,7 +17,7 @@
   # shell.nix 
   { repoRoot, inputs, pkgs, lib, system }:
   lib.iogx.mkShell {
-    tools.nixfmt.enable = true;
+    tools.nixfmt-classic.enable = true;
   }
   ```